Support for Big SSD
-Tested with up to 2TB disk size
-Added support for Delay mode:
    1.real - backwards compatible, same delay that was
      implemented until now
    2.emulated - no delay, execution of test continues normally,
      delay is only logged.
-All relevant variable in code were updated to the appropriate kind,
 size of variable was increased to accommodate larger disk, and
 signed/unsigned was changed depending on actual use of variable
-Calculation of ssd_stats: utilization and read/write speed was fixed
 to account for multiple rt_analyzers
-Added ssd_stats: occupied_pages, logical_write_count,
 block_erase_count, channel_switch_to_write, channel_switch_to_read
-Added mechanism for syncing to log without arbitrary time sleep
-Updated GC: Victim list is now sorted by valid_page_nb, accelerating
 GC algorithm for large disk sizes.
-Update legacy tests for new changes
-ssd_write_read test now has tests for checking all ssd_stats
 during test of read, write, mixed read/write and GC tests

+++ eVSSIM/CONFIG/vssim_config_manager.h
@@ -11,28 +11,28 @@
 
 /* SSD Configuration */
 extern uint64_t SECTOR_NB;
-extern uint32_t PAGE_NB;
+extern uint64_t PAGE_NB;
 extern uint32_t FLASH_NB;
-extern uint32_t BLOCK_NB;
+extern uint64_t BLOCK_NB;
 extern uint32_t CHANNEL_NB;
 extern uint32_t PLANES_PER_FLASH;
 
 extern uint32_t SECTORS_PER_PAGE;
-extern uint32_t PAGES_PER_FLASH;
-extern int64_t PAGES_IN_SSD;
+extern uint64_t PAGES_PER_FLASH;
+extern uint64_t PAGES_IN_SSD;
 
 extern uint32_t WAY_NB;
 
 /* Mapping Table */
 extern uint32_t DATA_BLOCK_NB;
-extern int64_t BLOCK_MAPPING_ENTRY_NB;		// added by js
+extern uint64_t BLOCK_MAPPING_ENTRY_NB;		// added by js
 
 #ifdef PAGE_MAP
-extern int64_t PAGE_MAPPING_ENTRY_NB; 		// added by js
-extern int64_t EACH_EMPTY_TABLE_ENTRY_NB;	// added by js
+extern uint64_t PAGE_MAPPING_ENTRY_NB; 		// added by js
+extern uint64_t EACH_EMPTY_TABLE_ENTRY_NB;	// added by js
 
-extern uint32_t EMPTY_TABLE_ENTRY_NB;
-extern uint32_t VICTIM_TABLE_ENTRY_NB;
+extern uint64_t EMPTY_TABLE_ENTRY_NB;
+extern uint64_t VICTIM_TABLE_ENTRY_NB;
 #endif
 
 /* NAND Flash Delay */
